# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "University Ecosystem API",
    "description": "REST API for university life management platform - schedules, news, events, notifications, and more.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/chats": {
      "get": {
        "tags": [
          "chats"
        ],
        "summary": "Get Chats",
        "description": "Get all chats for the current user with cursor-based pagination.\n\nReturns chats ordered by last message timestamp (newest first).\nUse the `next_cursor` from the response to fetch the next page.",
        "operationId": "get_chats_api_v1_chats_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Pagination cursor",
              "title": "Cursor"
            },
            "description": "Pagination cursor"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Number of chats to return",
              "default": 20,
              "title": "Limit"
            },
            "description": "Number of chats to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatsListOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "chats"
        ],
        "summary": "Create Chat",
        "description": "Create a new chat with a user. If a chat already exists, return it.",
        "operationId": "create_chat_api_v1_chats_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/chats/{chat_id}": {
      "get": {
        "tags": [
          "chats"
        ],
        "summary": "Get Chat",
        "description": "Get details for a specific chat.",
        "operationId": "get_chat_api_v1_chats__chat_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "chat_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Chat Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "chats"
        ],
        "summary": "Delete Chat",
        "description": "Delete a chat entirely for all participants (messages, attachments, links).",
        "operationId": "delete_chat_api_v1_chats__chat_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "chat_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Chat Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatMaintenanceResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/chats/{chat_id}/messages": {
      "get": {
        "tags": [
          "chats"
        ],
        "summary": "Get Messages",
        "description": "Get messages for a chat with cursor-based pagination.\n\nMessages are returned in ascending order (oldest first).\nUse the `next_cursor` from the response to fetch older messages.",
        "operationId": "get_messages_api_v1_chats__chat_id__messages_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "chat_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Chat Id"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Pagination cursor",
              "title": "Cursor"
            },
            "description": "Pagination cursor"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Number of messages to return",
              "default": 50,
              "title": "Limit"
            },
            "description": "Number of messages to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessagesListOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "chats"
        ],
        "summary": "Send Message",
        "description": "Send a message to a chat.\n\nThe message is saved to the database and all chat participants\nare notified via WebSocket in real-time.",
        "operationId": "send_message_api_v1_chats__chat_id__messages_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "chat_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Chat Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_send_message_api_v1_chats__chat_id__messages_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/chats/{chat_id}/read": {
      "post": {
        "tags": [
          "chats"
        ],
        "summary": "Mark Read",
        "description": "Mark all messages in a chat as read.",
        "operationId": "mark_read_api_v1_chats__chat_id__read_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "chat_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Chat Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/chats/{chat_id}/clear": {
      "post": {
        "tags": [
          "chats"
        ],
        "summary": "Clear Chat History",
        "description": "Remove all messages (and attachments) from a chat for its participants.",
        "operationId": "clear_chat_history_api_v1_chats__chat_id__clear_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "chat_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Chat Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatMaintenanceResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AttachmentResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "file_type": {
            "type": "string",
            "title": "File Type"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          }
        },
        "type": "object",
        "required": [
          "id",
          "url",
          "file_type",
          "filename",
          "size"
        ],
        "title": "AttachmentResponse"
      },
...
      "Body_send_message_api_v1_chats__chat_id__messages_post": {
        "properties": {
          "content": {
            "type": "string",
            "title": "Content",
            "default": ""
          },
          "files": {
            "items": {
              "type": "string",
              "format": "binary"
            },
            "type": "array",
            "title": "Files",
            "default": []
          }
        },
        "type": "object",
        "title": "Body_send_message_api_v1_chats__chat_id__messages_post"
      },
...
      "ChatCreate": {
        "properties": {
          "participant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Participant Id"
          }
        },
        "type": "object",
        "required": [
          "participant_id"
        ],
        "title": "ChatCreate"
      },
      "ChatMaintenanceResult": {
        "properties": {
          "chat_id": {
            "type": "string",
            "format": "uuid",
            "title": "Chat Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "deleted_messages": {
            "type": "integer",
            "title": "Deleted Messages",
            "default": 0
          },
          "deleted_attachments": {
            "type": "integer",
            "title": "Deleted Attachments",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "chat_id",
          "status"
        ],
        "title": "ChatMaintenanceResult",
        "description": "Represents the result of a maintenance operation on a chat."
      },
      "ChatParticipant": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "is_active"
        ],
        "title": "ChatParticipant"
      },
      "ChatResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "participants": {
            "items": {
              "$ref": "#/components/schemas/ChatParticipant"
            },
            "type": "array",
            "title": "Participants"
          },
          "last_message": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MessageResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "unread_count": {
            "type": "integer",
            "title": "Unread Count",
            "default": 0
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "presence": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/PresenceStatus"
                },
                "propertyNames": {
                  "format": "uuid"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Presence"
          }
        },
        "type": "object",
        "required": [
          "id",
          "participants",
          "created_at",
          "updated_at"
        ],
        "title": "ChatResponse"
      },
      "ChatsListOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ChatResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More",
            "default": false
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "ChatsListOut",
        "description": "Paginated list of chats."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MessageResponse": {
        "properties": {
          "content": {
            "type": "string",
            "title": "Content"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "chat_id": {
            "type": "string",
            "format": "uuid",
            "title": "Chat Id"
          },
          "sender_id": {
            "type": "string",
            "format": "uuid",
            "title": "Sender Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "read_status": {
            "type": "boolean",
            "title": "Read Status"
          },
          "sender": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ChatParticipant"
              },
              {
                "type": "null"
              }
            ]
          },
          "sender_presence": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PresenceStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "attachments": {
            "items": {
              "$ref": "#/components/schemas/AttachmentResponse"
            },
            "type": "array",
            "title": "Attachments",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "content",
          "id",
          "chat_id",
          "sender_id",
          "created_at",
          "read_status"
        ],
        "title": "MessageResponse"
      },
      "MessagesListOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MessageResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More",
            "default": false
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "MessagesListOut",
        "description": "Paginated list of messages."
      },
...
      "PresenceStatus": {
        "properties": {
          "active": {
            "type": "boolean",
            "title": "Active",
            "default": false
          },
          "last_seen_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Seen At"
          }
        },
        "type": "object",
        "title": "PresenceStatus",
        "description": "Represents a participant's presence state."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/auth/login"
          }
        }
      }
    }
  }
}